Make request_slow_test robust to initial db state

+++ server_code/request_slow_test.py
@@ -90,7 +90,8 @@ class TestPropRequestProp(unittest.TestCase):
 class TestRequestGateway(unittest.TestCase):
   def setUp(self):
     self.request_records_saved = []
-    self.request_rows_created = []
+    self.are_rows_to_delete = False
+    self.test_start_dt = sm.now()
     self._email_send = n.email_send
     n.email_send = Mock()
     self._send_sms = n.send_sms
@@ -103,8 +104,8 @@ class TestRequestGateway(unittest.TestCase):
     request = next(rs.prop_to_requests(prop, with_users=[rt.admin_id]))
     request_record = rg.RequestRecord(request)
     self.assertFalse(request_record.record_id)
-    request_record.save()
     self.request_records_saved.append(request_record)
+    request_record.save()
     self.assertTrue(request_record.record_id)
     saved_requests = [r for r in rg.requests_by_user(USER2) 
                       if r.request_id == request_record.record_id]
@@ -128,12 +129,12 @@ class TestRequestGateway(unittest.TestCase):
 
   def test_cancel_request(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=or_group_id0))
     requests = list(rg.current_requests(records=False))
     self.assertEqual(len(requests), 1)
     ri.cancel_request(USER3, requests[0].request_id)
     edited_requests = list(rg.current_requests(records=False))
     self.assertEqual(len(edited_requests), 0)
+    self.are_rows_to_delete = True
   
   def test_edit_request(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
@@ -145,7 +146,6 @@ class TestRequestGateway(unittest.TestCase):
     requests[0].max_size = 3
     _prop = list(ri.requests_to_props(requests, USER3))[0]
     or_group_id1 = edit_request(USER3, _prop)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=or_group_id0))
     edited_requests = list(rg.current_requests(records=False))
     request_id1 = edited_requests[0].request_id
     self.assertEqual(or_group_id0, or_group_id1)
@@ -153,10 +153,10 @@ class TestRequestGateway(unittest.TestCase):
     self.assertEqual(len(edited_requests), 1)
     self.assertEqual(edited_requests[0].or_group_id, or_group_id1)
     self.assertEqual(edited_requests[0].max_size, 3)
+    self.are_rows_to_delete = True
 
   def test_edit_request_change_eligible(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=or_group_id0))
     requests = list(rg.current_requests(records=False))
     self.assertEqual(len(requests), 1)
     request_id0 = requests[0].request_id
@@ -172,35 +172,34 @@ class TestRequestGateway(unittest.TestCase):
     self.assertEqual(request_id0, request_id1)
     self.assertEqual(len(edited_requests), 1)
     self.assertEqual(edited_requests[0].or_group_id, or_group_id1)
+    self.are_rows_to_delete = True
   
   def test_visible_requests(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
     or_group_id1 = add_request(ADMIN, rt.prop_uA_2to2_in1hr)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=q.any_of(or_group_id0, or_group_id1)))
     visible_requests = ri.current_visible_requests(USER2, list(rg.current_requests(records=True)))
     self.assertEqual(len(visible_requests), 1)
     self.assertEqual(visible_requests[0].or_group_id, or_group_id1)
+    self.are_rows_to_delete = True
 
   def test_visible_request_with(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=q.any_of(or_group_id0)))
     request = next(rs.prop_to_requests(rt.prop_uA_2to2_in1hr, with_users=[rt.u2.user_id]))
     request_record = rg.RequestRecord(request)
     request_record.save()
-    self.request_records_saved.append(request_record)
     visible_requests = ri.current_visible_requests(USER2, list(rg.current_requests(records=True)))
     self.assertEqual(len(visible_requests), 1)
     self.assertEqual(visible_requests[0].or_group_id, request.or_group_id)
+    self.are_rows_to_delete = True
 
   def test_not_visible_request_with(self):
     or_group_id0 = add_request(USER3, rt.prop_u3_3to10_in1hr)
-    self.request_rows_created.extend(app_tables.requests.search(or_group_id=q.any_of(or_group_id0)))
     request = next(rs.prop_to_requests(rt.prop_uA_2to2_in1hr, with_users=[rt.u3.user_id]))
     request_record = rg.RequestRecord(request)
     request_record.save()
-    self.request_records_saved.append(request_record)
     visible_requests = ri.current_visible_requests(USER2, list(rg.current_requests(records=True)))
     self.assertFalse(visible_requests)
+    self.are_rows_to_delete = True
   
   def tearDown(self):
     n.email_send = self._email_send
@@ -208,5 +207,7 @@ class TestRequestGateway(unittest.TestCase):
     ri.RequestManager.notify_edit = self._notify_edit
     for rr in self.request_records_saved:
       rr._row.delete()
-    for row in self.request_rows_created:
-      row.delete()
+    if self.are_rows_to_delete:
+      rows_created = app_tables.requests.search(create_dt=q.greater_than_or_equal_to(self.test_start_dt))
+      for row in rows_created:
+        row.delete()
